fix: rename files and folders
