+++ src/stacks-hierarchy/RootStack.tsx
@@ -54,7 +54,7 @@ import {
   Root_ParkingViolationsPhotoScreen,
   Root_ParkingViolationsQrScreen,
   Root_ParkingViolationsSelectScreen,
-} from '@atb/stacks-hierarchy/Root_ParkingViolationsReporting';
+} from '@atb/stacks-hierarchy/Root_ScooterHelp';
 import {Root_NotificationPermissionScreen} from '@atb/stacks-hierarchy/Root_NotificationPermissionScreen';
 import {Root_LocationWhenInUsePermissionScreen} from '@atb/stacks-hierarchy/Root_LocationWhenInUsePermissionScreen';
 import {useBeaconsContext} from '@atb/beacons/BeaconsContext';
@@ -66,9 +66,9 @@ import {useRegisterIntercomUser} from '@atb/chat/use-register-intercom-user';
 import {useRemoteConfigContext} from '@atb/RemoteConfigContext';
 import {ForceUpdateScreen} from '@atb/force-update-screen';
 import {compareVersion} from '@atb/utils/compare-version';
-import {Root_ScooterHelpScreen} from './Root_ParkingViolationsReporting/Root_ScooterHelpScreen';
-import {Root_ScooterContactOperatorScreen} from './Root_ParkingViolationsReporting/Root_ScooterContactOperatorScreen';
-import {Root_ScooterOperatorContactConfirmationScreen} from './Root_ParkingViolationsReporting/Root_ScooterOperatorContactConfirmationScreen';
+import {Root_ScooterHelpScreen} from './Root_ScooterHelp/Root_ScooterHelpScreen';
+import {Root_ContactScooterOperatorScreen} from './Root_ScooterHelp/Root_ContactScooterOperatorScreen';
+import {Root_ContactScooterOperatorConfirmationScreen} from './Root_ScooterHelp/Root_ContactScooterOperatorConfirmationScreen';
 
 type ResultState = PartialState<NavigationState> & {
   state?: ResultState;
@@ -395,12 +395,12 @@ export const RootStack = () => {
                 component={Root_ScooterHelpScreen}
               />
               <Stack.Screen
-                name="Root_ScooterContactOperatorScreen"
-                component={Root_ScooterContactOperatorScreen}
+                name="Root_ContactScooterOperatorScreen"
+                component={Root_ContactScooterOperatorScreen}
               />
               <Stack.Screen
-                name="Root_ScooterOperatorContactConfirmationScreen"
-                component={Root_ScooterOperatorContactConfirmationScreen}
+                name="Root_ContactScooterOperatorConfirmationScreen"
+                component={Root_ContactScooterOperatorConfirmationScreen}
               />
               <Stack.Screen
                 name="Root_ParkingViolationsSelectScreen"
