+++ src/stacks-hierarchy/Root_ScooterHelp/Root_ContactScooterOperatorConfirmationScreen.tsx
@@ -1,16 +1,16 @@
 import {RootStackScreenProps} from '@atb/stacks-hierarchy';
 import {useTranslation} from '@atb/translations';
 import {useAnalyticsContext} from '@atb/analytics';
-import {ConfirmationScreen} from '@atb/scooter-help/components/ConfirmationScreen';
+import {ConfirmationScreen} from '@atb/stacks-hierarchy/Root_ScooterHelp/components/ConfirmationScreen';
 import {ScooterOperatorContactTexts} from '@atb/translations/screens/ScooterOperatorContact';
 
-export type ConfirmationScreenProps =
-  RootStackScreenProps<'Root_ScooterOperatorContactConfirmationScreen'>;
+export type ContactScooterOperatorConfirmationScreenProps =
+  RootStackScreenProps<'Root_ContactScooterOperatorConfirmationScreen'>;
 
-export const Root_ScooterOperatorContactConfirmationScreen = ({
+export const Root_ContactScooterOperatorConfirmationScreen = ({
   navigation,
   route: {params},
-}: ConfirmationScreenProps) => {
+}: ContactScooterOperatorConfirmationScreenProps) => {
   const {t} = useTranslation();
   const analytics = useAnalyticsContext();
