+++ src/stacks-hierarchy/Root_ScooterHelp/Root_ParkingViolationsConfirmationScreen.tsx
@@ -2,7 +2,7 @@ import {RootStackScreenProps} from '@atb/stacks-hierarchy';
 import {useTranslation} from '@atb/translations';
 import {ParkingViolationTexts} from '@atb/translations/screens/ParkingViolations';
 import {useAnalyticsContext} from '@atb/analytics';
-import {ConfirmationScreen} from '@atb/scooter-help/components/ConfirmationScreen';
+import {ConfirmationScreen} from '@atb/stacks-hierarchy/Root_ScooterHelp/components/ConfirmationScreen';
 
 export type ConfirmationScreenProps =
   RootStackScreenProps<'Root_ParkingViolationsConfirmationScreen'>;
