+++ src/stacks-hierarchy/Root_ScooterHelp/bottom-sheets/ImageConfirmationBottomSheet.tsx
