+++ src/stacks-hierarchy/Root_ScooterHelp/components/ProviderLogo.tsx
