+++ src/stacks-hierarchy/navigation-types.ts
@@ -74,12 +74,12 @@ type Root_ScooterHelpScreenParams = {
   vehicleId: string;
 };
 
-type Root_ScooterContactOperatorScreenParams = {
+type Root_ContactScooterOperatorScreenParams = {
   vehicleId: string;
   operatorId: string;
 };
 
-type Root_ScooterOperatorContactConfirmationScreenParams = {
+type Root_ContactScooterOperatorConfirmationScreenParams = {
   operatorName: string;
 };
 
@@ -134,8 +134,8 @@ export type RootStackParamList = StackParams<{
   Root_ParkingViolationsQrScreen: Root_ParkingViolationsQrParams;
   Root_ParkingViolationsConfirmationScreen: Root_ParkingViolationsConfirmationParams;
   Root_ScooterHelpScreen: Root_ScooterHelpScreenParams;
-  Root_ScooterContactOperatorScreen: Root_ScooterContactOperatorScreenParams;
-  Root_ScooterOperatorContactConfirmationScreen: Root_ScooterOperatorContactConfirmationScreenParams;
+  Root_ContactScooterOperatorScreen: Root_ContactScooterOperatorScreenParams;
+  Root_ContactScooterOperatorConfirmationScreen: Root_ContactScooterOperatorConfirmationScreenParams;
   Root_NotificationPermissionScreen: undefined;
   Root_LocationWhenInUsePermissionScreen: undefined;
   Root_ChooseTicketRecipientScreen: Root_ChooseTicketRecipientScreenParams;
